This is it, chief
This is as tried and tested as I can get this thing right now.
